+++ src/config/filterizrState.ts
@@ -1,7 +1,14 @@
+interface filterizrState {
+  IDLE: 'IDLE';
+  FILTERING: 'FILTERING';
+  SORTING: 'SORTING';
+  SHUFFLING: 'SHUFFLING';
+}
+
 /**
  * Possible animation states for Filterizr
  */
-export const FILTERIZR_STATE = {
+export const FILTERIZR_STATE: filterizrState = {
   IDLE: 'IDLE',
   FILTERING: 'FILTERING',
   SORTING: 'SORTING',