+++ src/config/index.ts
@@ -1,3 +1,2 @@
 export { FILTERIZR_STATE } from './filterizrState';
-export { TRANSITION_END_EVENTS } from './transitionEndEvents';
 export { cssEasingValuesRegexp } from './cssEasingValuesRegexp';