+++ src/types/index.ts
@@ -1,1 +1,2 @@
+export type FilterizrState = 'IDLE' | 'FILTERING' | 'SORTING' | 'SHUFFLING';
 export type Filter = string | string[];