+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Change column calculation logic to use `Math.floor` instead of `Math.round` (fixes overflowing elements at the grid's end)
 - Add new public method `removeItem` that removes an item from the grid
 - Add a built-in loading spinner animation controlled through the `spinner` option in options.
+- Fix timing of callbacks
 
 ## Version 2.1.1